# perf, skip some syscalls